mock in general modules from other yast2 modules
